Wait on vAppConfig during async create
This patch updates the behavior of the vAppConfig, LinuxPrep, and
Sysprep bootstrap providers to wait on the VM to have a non-nil
config.vAppConfig property when it is expected to exist.

Async create means we are no longer falling through to the update
logic directly after create, instead, relying on the reconciler to
be reentrant. That all works fine, but the problem is vpxd is sending
signal on the property collector that the VM is ready prior to its
config.vAppConfig property being set.

VM Op receives this signal and proceeds to reconcile the VM. Since
previously we did not wait on an expected vAppConfig, we proceeded
to configure and power on the VM. Once it is powered on, any customization
that dependended upon the vAppConfig properties being set prior to
first boot will no longer work.

This patch returns an error from GetOVFVAppConfigForConfigSpec if the
VM has a nil config.vAppConfig.