Misc VirtualMachineSetResourcePolicy improvements
The child Folder and ResourcePool names are optional fields but we
were not checking for this.

When getting both the Resource Pool and Folder MoIDs of a namespace,
don't always return the last FolderMoID on the Zone or AZ. The Folder
is VC scoped and we should not ever have a Zone or AZ without it,
return the value if at least one of the Zone/AZ has it set.

Use the v1a2+ field names in the validation webhook error messages