Add VirtualMachineWebConsoleRequest in web-console-validator (#820)
This PR adds a missing check for VirtualMachineWebConsoleRequest in the web-console-validator server. This CR was introduced in v1alpha2 (renamed from WebConsoleRequest), so we need to handle both CRs to support all API versions.

While here, doing some code refactor and moving VirtualMachineWebConsoleRequest controller and webhook from the v1alpha2 folder to the root folder to avoid confusion (since we're now in v1alpha3).