Fix potential VM validation webhook NPE for Spec.Network
If the VM Spec.Network was changed to/from nil, the validation webhook
would NPE. I think the stems from late in v1a2 we decided to make most
of the top level Spec/Status fields pointers and this was just missed.

Add a few basic update tests for VM Spec.Network.