Do not CBLO if VM watcher service fails
This patch updates the behavior of the VM watcher service to no
longer return an error if there is an issue starting. While this
is the correct behavior, it also results in CBLO for the VM Op
pod if the reason for the error is external and not resolved.

For example, if a user configures the host name of vCenter to be
invalid, the VM Op pod will be in CBLO until this is resolved.
Again, this is technically correct, and resolves itself once the
user fixes the misconfigured setting.

However, the VM Op pod becoming unavailable on a single-node
Supervisor also means no webhooks are running, and this would be
bad.