Introduce a new annotation to identify failed over VMs (#829)
We need a new annotation for failed over VMs so we can skip some
webhook checks.  This is different from imported, or registered VMs
since we will need to relax additional webhook checks such as allowing
for network manipulation.

Remove the test failover annotation since that is no longer needed.
Also, rename the registered annotation to reflect that it is a restore
operation.