Refactor chanErr consumption in VM controller
This patch changes the way the error channel from the async create
workflow is consumed in the VirtualMachine controller.

Previously only a single error was assumed to be returned, but since
this is a channel, the consumer should range over it, accounting for
the possible vim and k8s error.