Reconcile VM SetResourcePolicy because of Zone add
When a Zone is added, the VirtualMachineSetResourcePolicies in the
namespace need to be reconciled because we need to create the child
ResourcePool and ClusterModules for that new CCR. Otherwise, the
new Zone won't be ready to use until the next resync.

While at the moment we only really care about when a Zone is added
just watch for all events since we eventually will need to handle
Zone removal.