Remove CL Item ContentVersion short circuit
To avoid repeated CL Item sync tasks, we'll compare the item and image
cached content versions, and skip fetching the OVF envelope if they
match. But since we've added fields to the image Status (like Disks[])
in v1a3, those new fields won't be updated until the item changes.

The big hammer for this is to just remove the check instead of including
another internal version in the cached version. Short term, this will
cause an increase in the number of sync tasks, but the existing OVF
cache will reduce that somewhat. There is a larger WIP in this area that
will fix this bug without the increase in sync tasks.