Fix empty VMI Status.Disks[] after upgrade
The VMI Status.Disks[] field was added in v1a3 but since the cached
content version will be the same, that field won't be populated.

Just a quick fix just checking for the present of disks. There is a WIP
branch in this area that will fix it in a better way.