Consistently use the default config for integration tests
Also fix the feature state for the ConfigMap capabilities controller.
It is only used when SVAsyncUpgrade is false but the setup was setting
it to true. The fallout from this is that the WorkloadDomainIsolation
feature is not via the capability when async upgrade is false so remove
that from the integration tests.