VM Watcher Service improvements
When building the initial container MoID list, include the IDs so
that we can match the Add() and Remove() of the same ID. We had a
bug here in that if the Zone was initially added to a NS when at
start up time, any later removal of the Zone from a NS would remove
the watch of that Zone across all namespaces. This allows the watcher
Add() and Remove() to be idempotent which is needed for some infra
Zone controller described later.

Tweak how the infra Zone controller handles the finalizer. The VC
watcher itself is ephemeral since it will go away when the VC client
is closed. But we need the Zone finalizer so that we can remove it
from the watcher when the Zone goes away. Therefore, call Add() or
Remove() on the Zone regardless if the finalizer was already present
or not.

Better handle a Zone with a stale FolderMoID: if a folder does not
exist on VC, the watcher will fail to start. We filter out invalid
container MoIDs to the watcher; the Zone controller will later keep
retrying to add that FolderMoID to the watcher.

Much of change is fallout from 14c1ac32b when we changed the VM
watcher service to not os.Exit() whenever the watcher could not
be started.