Background reconcile awareness support
This patch adds support to the VirtualMachine controller for
background reconciliation awareness. Rather, the controller is
now aware if a VM is already being reconciled in a background
thread due to an async operation and will not allow the VM to be
reconciled again until the async operation is complete.