Remove FSS WorkloadDomainIsolation
The feature WorkloadDomainIsolation has been enabled and thus its
feature state switch should be removed, the enabled logic acting as
the new, default, and only logic.

Please note, the updated placement logic is always true for
non-vSphere Kubernetes Service (VKS) workloads. For VKS nodes, the
placement logic is determined by the value of the
WorkloadDomainIsolation capability.

Please note, the env var for the FSS is retained in the deployment
YAML as "true" so as not to break testing that relies on reading this
value.