Print memory statistics
This package adds support for emitting the results of a
periodic call to runtime.ReadMemStats, enabling easily
discovering the memory usage of the containers from the
perspective of the Go runtime.

The environment variable MEM_STATS_PERIOD may be set to a
valid time.Duration string to configure the interval, and
it defaults to 10m.

The memory usage is emitted as soon as main.go begins, ex.:

        I0121 09:38:11.853666   91609 mem.go:30] "MemStats" now=1737473891853622000 buckHashSys=1449862 debugGC=false enableGC=true frees=9602 gcCPUFraction=0.013948351398451445 gcSys=3138792 heapAlloc=4623544 heapIdle=1736704 heapInUse=6127616 heapObjects=18663 heapReleased=1703936 heapSys=7864320 lastGC=1737473891832972000 lookups=0 mCacheInUse=12000 mSpanInUse=126560 mSpanSys=130560 mallocs=28265 nextGC=6347392 numForcedGC=0 numGC=2 otherSys=1181098 pauseTotalNs=77666 stackInUse=524288 stackSys=524288 sys=14304520 totalAlloc=6120904

Please refer to runtime.MemStats,
https://pkg.go.dev/runtime#MemStats,
for more information on the keys/values.