Include UUID of disk backing in disk backup (#864)
This change modifies the disk backups to also include disk backing
UUIDs.  This allows us to look up and identify disks using UUIDs to
differentiate classic disks from disks that should be FCDs.

For now, the disk UUIDs are only used in failover workflows since
vSphere Replication guarantees that the disks will be restored with
the same UUID as the primary/source.