Return taskInfo from successful VM Reconfigure task
This can contain useful info even when the task succeeds. We
currently only ever care about the taskInfo when we also care
if the task fails, so this was otherwise working.

This resource VirtualMachine package is super old, and only
ever tested indirectly, and my eventual plan is to either
remove it entirely or move it functionality into the
virtualmachine package so just quick fix this now so it
doesn't trip up anybody else later.