Remove UpdateConfigSpecFirmware()
This is from the early days when we could only update the VM's
Firmware - and most things for that matter - after the VM was
created since Content Library did not support creating from a
ConfigSpec. Since now CL supports that, and some time back I
added support in CreateConfigSpec() for the firmware annotation,
so the VM is just created with the specified firmware.

Changing the firmware of existing VM is generally not going to
result in a booting VM, so at the moment I didn't implement
similar support in resize.