Do deployment rollout when capabilities change
This patch changes the behavior of how the pod is exited when the
capabilities have changed. Instead of all the replicas exiting at
the same time, the exits are managed by causing the pod's deployment
to rollout a new set of pods.

This ensures that at least one pod is always online to respond to
webhook requests.

This patch also ensures the capabilities controllers are run only
on the leader.