Use UnsafeDisableDeepCopy when possible
This patch updates many K8s client.List calls to use
the UnsafeDisableDeepCopy option where it is possible
to do so. This is done with caution, and only in places
where the result is self-contained and there is little
chance of mutating the objects from the cache. There is
also a warning placed on every location where this option
is used.

This patch improves performance.