Defaults for ConfigSpec CPU/mem alloc
This patch ensures there are defaults for the ConfigSpec CPU/mem
allocation fields for limits and shares.

The documentation for a ResourceAllocation object states that all (VM)
fields must be set for certain APIs, including ImportVApp. A part of
that API is used by the vpxd placement engine to construct fake VMs in
order to support assignable hardware. Therefore:

CPU allocation:
- Ensure limit is set to -1 if there is a reservation.

Memory allocation:
- Ensure shares is not nil.
- Ensure limit is set to -1 if there is a reservation.