Remove FSS ISO Support (#904)
This PR removes the ISO Support FSS related code since it's been enabled on main for a while. The env var for the FSS is retained in the deployment YAML as "true" to avoid breaking internal tests that rely on reading this value.

Ran gobuild internally and verified VMOP deployment from that build.