Zone placement group and org
This patch introduces the concepts of a Zone Placement Group (ZPG)
and Zone Placement Org (ZPO). VM workloads in a ZPG are affined to
a single zone with the zone discovered by using DRS to do placement
for a single member of the ZPG. ZPGs in a ZPO are automatically
spread across available zones. If the number of ZPGs in a ZPO exceeds
the number of available zones then the remaining ZPGs are placed
non-deterministically.

The label "vmoperator.vmware.com/zone-placement-group" is used to
define a ZPG. The label "vmoperator.vmware.com/zone-placement-org"
is used to define a ZPO.

VMs may be members of ZPGs without also being members of ZPOs.
However, VMs MUST be members of ZPGs in order to participate in a
ZPO.