Do not list all StoragePolicyQuotas during SPQ reconcile
When reconciling a SQP, we would later end up listing all the
SQP in the namespace to determine what storage classes are
associated with the policy quota. We don't need to list all
the SQPs since each one will be reconciled.

While we can't - or should not - have the same storage policy
ID for multiple SQP in a namespace, but if we could this cause
a failure on the reconcile of later SQPs since we'd try to set
multiple controller owner references.