Add option to not configure gateway on interfaces
Allow setting the Interface Gateway{4,6} to "None" to indicate to not
set the gateway that is provided by the underlying network provider in
the network configuration. This can be useful when a VM is configured
with multiple interfaces.

Remove an incorrect documentation note about the gateway: it is
allowed to manually provide IP addresses without a gateway.