Revert v1a3 vm.status.storage.usage to 9.x
This patch updates the VM API status.storage.usage data
structure to be what it used to be, and what it shipped as
in vC 9.0.

The intent was to ship vC 9.0 with the same status.storage.usage
structure now seen in v1alpha4, but there were some reverts, and
those changes were discarded.