Support sysprep joinDomainMode
This patch adds support for specifying the mode used to join a
system to Active Directory. The field spec.bootstrap.joinDomainMode
may be set to either Create (the default) or Wait. The latter value
causes the VM Op pause annotation to be applied to the VM, halting
reconciliation until an external control plane creates the Computer
object in AD and removes the pause annotation from the VM.