Require VM ResourcePolicyName when specifying a cluster module
The VM SetResourcePolicy contains the cluster module info, so if the VM
has the special cluster module annotation, enforce that the resource
policy is set.  Post create, we only need the VM's resource policy to
set the VM cluster module. We really only need if it is not already set
but that is not easy to check as this point.