Set VM as controller owner ref on network interfaces
This was an oversight from the original code. Do not attempt to use an
existing network interface that is currently controller owned by another
VM. For any existing situations where the interface has an controller
owner that is not a VM add just an owner ref like before to keep things
working.