FastDeploy: Use Datastore method to check if cache files exist
QueryVirtualDiskUuid only works for .vmdk, we need to check if other types exist, such as .nvram

Originally planned to use HostDatastoreBrowser.SearchDatastore via govmomi's
object.Datastore.Stat method, however:
- Would need to refactor so object.Datastore is plumbed through,
  instead of string DatastorePath: "[DatastoreName] dir/file.ext"
- Considered using govmomi object.DatastoreFile.Stat, but has the same plumbing issue
- The Datastore HEAD method can use DatastorePath and avoids creating a Task / WaitForUpdates
- Both methods require the same Datastore.Browse permission

The getDatacenters helper method was collecting Datacenter.Name, which validated the moid,
but Name property was not used. Changed to get the InventoryPath which also validates the moid and
is used to construct the Datastore HEAD request URL. Note that while setting:
object.Datastore.InventoryPath = Datacenter.Name
would work is most environments, the full path is needed if Datacenter is in a sub Folder with the RootFolder.