Introduce VirtualMachineSnapshot API
This change introduces the APIs to take a snapshot of a VM Service VM.
A new custom resource definition is added that contains the
desired specifications and the observed status of the snapshot.

VirtualMachine custom resource definition has been modified to enable
workflows for creating, deleting a snapshot as well as reverting a VM
to a snapshot.