Support skip-validation annotation
This patch adds support for an annotation that works across all
VM Operator objects and allows a privileged user to indicate that
the validation webhooks should be skipped for the object to which
said annotation is applied.

Please note, this annotation also results in bypassing the quota
check for an object.