Prefer cloudinit local-ipv4/6 for VM's Primary IPs
The cloudinit VMware datasource sets these values in the ExtraConfig, so
use those fields if set as the VM's primary IPv4/6 over what is reported
by VM Tools. Apply the same validity checks on EC fields as what we were
already doing.

Since we were already both walking and making a map of the ExtraConfig
convert it to a map once and pass that down.