Report disk capacity for usage
This patch updates the way VM Op reports VM disk usage against
the namespace's storage quota. Previously the usage was reported
via the actual usage on disk, i.e. even if a disk was provisioned
with a capacity of 50GiB, and it was only actually taking up 20GiB
on disk, then only the 20GiB would be reported against quota.
Now the disk's potential usage, its total capacity, is reported for
quota accounting.

However, a user can still determine a disk's usage vs. capacity as
well as the same for the entire VM via status.storage. This field
has been updated such that:

* `status.storage.used.other` is the total amount of storage used
  by all of the VM's non disks.
* `status.storage.used.disks` is the total amount of storage used
  by all of the VM's non-PVC disks.
* `status.storage.requested.disks` is the total amount of storage
  requested by all of the VM's non-PVC disks.
* `status.storage.total` is a sum of `status.storage.used.other`
  and `status.storage.requested.disks`.

The value of `status.storage.total` is what is reported against
the namespace's storage quota.