Support Supervisor+VPC DNS lookups in guests
This patch adds support for propagating the Supervisor's DNS into
guests via the Supervisor kube-system/kube-dns-lb service, which
exposes the Supervisor DNS over an IP address visible to workload
networks.

Please note this currently only works on the VPC Supervisor network
topology. This is because only that topology creates the service
kube-system/kube-dns-lb to surface the Supervisor's DNS to workload
networks.