Skip delete platform resource annotation
This patch adds support for a privileged annotation that may be
used to skip the deletion of an underlying platform resource when
the corresponding Kubernetes object is deleted.

This supports use cases such as rollback of import where the underlying
VM should not be deleted even though the associated Kubernetes object
is.