Graceful VM dir cleanup / crypto+linked
This patch fixes an issue with how the VM home dir is cleaned up
upon a failed fast deploy of a VM. If the VM home dir is already
gone, the old code failed with a FileNotFound fault. Now that
fault is ignored.

This patch also defaults VMs to use direct mode with Fast Deploy
if the VM is using an encrypted storage policy.