Introduce spec.class for VirtualMachines (#1013)
This change introduces spec.class for VirtualMachine resources.  This
reference points to a VirtualMachineClassInstance resource.  This
field, together with spec.className uniquely represent the
hardware configuration of a VM.

The spec.className and spec.class fields always point to the same
underlying VM class.

If only one of these fields is specified, VM mutation webhook
populates the other.  If both are specified, the VM validation webhook
ensures that the references are valid.

A resize operation can be triggered by any of the following methods:
- Specifying a different class in spec.className field
- Specifying a new instance of the same class