Reconcile VM schema upgrade
This patch provides an officially supported path for upgrading a
VM Op VM's schema to match the latest version supported by VM Op.

Importantly, this change also disallows all updates to the VM Op
VM object until such time the VM object is updated. This is
critical as it prevents any changes that might interrupt or be
in conflict with the changes occurring as part of the upgrade.