Add VirtualMachineClassInstance reconciliation (#1015)
Reconcile VM classes to create VM class instance

An instance is created with an annotation containing the hash of the ConfigSpec. Additionally, the instance object also has an OwnerReference to the corresponding class. The controller ensures that at a time, only the instance mapping to the latest ConfigSpec hash is marked as active. The active / inactive property is represented by a label to allow clients to query efficiently.