Drop the Status sub-resource from payloads during backup (#1027)
Currently, when generating the backup manifest for a VM, or additional
resources, we include the full object - which also includes the Status
of the object.

The status of a resource is not useful during restore or failover
anyway.  On the contrary, this increases the size of information
stored in the VM's ExtraConfig, and thus, increasing the storage
consumption of the VM.  Additionally, since Status is of significant
size, VM operator has to spend extra CPU cycles and use memory to
construct this payload.

This change drops the Status from all payloads stored as part of
backup.