Mutating and validating webhooks for spec.class and spec.className (#1023)
This change introduces the webhook mutations and validations for
spec.className and spec.class of a VM.

The mutating webhook changes include:
- If spec.class is specified, but spec.className is omitted, the
mutating webhook populates the spec.className with the class that owns
the instance
- If spec.className is specified but spec.class is omitted, the
mutating webhook picks the active instance and populates the
spec.class with the instance.
- If both are specified, then mutating webhook is a pass-through.  The
validating webhook will validate the request.
- If neither of them are specified, then mutating webhook is a
pass-through.

The validating webhook changes include:
- If spec.className and spec.class both have been specified, they must
point to the same underlying class.

Additionally, the mutating webhook also applies an annotation
containing the last resized annotation name so later in the resize
workflow we can accurately evaluate if a resize has been requested.
This is the same behavior with className.