Handle network interface changes when VM is in powered off state
The VM's network interfaces can be modified when the VM is in the
off steady state, so reconcile any differences during that state
as well. This isn't strictly required since we will also do this
when the VM is in the transitioning to on state, but it is nice to
reflect changes before that.

Improve the checks the VM e2e tests to better assert that the
Ethernet device is that of what we expect by asserting the
External ID and MAC address.