Standardized controller reconcile logging
Generally, we were using different loggers for the FooContext and
the logger that is obtained from the reconcile context. This would
lead to different logging depending on which one was being used.

Add a LogConstructor that makes the controller-runtime logger less
verbose and more like what we have been logging in our FooContext
loggers.

Add util.FromContextOrDefault() function to use in most places in
lieu of logr.FromContextOrDiscard() just in case there is a situation
where the context provided doesn't include a logger.

Use the first 8 char of the controller-runtime reconcile ID as part of
the ID passed along to VC to easily correlate logs.

Remove the global VM Provider logger