Add support for reverting a VM to a given snapshot (#1041)
This change adds the capability to revert a VM from a given snapshot.
This can be done by the user by modifying the currentSnapshot field of
the VM's spec to point to the desired snapshot.

Along with reverting the VM to the given snapshot, we also roll back
the VM's Kubernetes state (Spec, Labels, Annotations) to the state
that was recorded in the Snapshot.  This is done by leveraging the
state recorded for backup / restore.

Other changes include:
- Update the create snapshot workflow to be conditioned on the VM
snapshot capability.  Also move the tests to be conditioned on the
capability.  Added a test to ensure that if the capability is not
enabled, no snapshots are created on the VM.
- A new annotation is added on the VM to indicate when a VM is being
reverted from a snapshot.  It is possible that the VC VM is
successfully reverted, but the VM's spec/metadata reverting fails.  In
that case, we would want to indicate to the user that this VM is _in_
restore from snapshot phase.  In a followup change, we will modify the
VM reconcile workflow to bail out if the VM is being reverted.
- Add a method to update the status of the VM with the current
snapshot details.
- A couple of helper methods to find the snapshot from the snapshot
tree of the VM.