Drive create snapshot workflow only through the custom resource
Initially, we were driving the snapshot workflow through the snapshot
custom resource modifying the VM's spec.currentSnapshot field -- which
will then be reconciled by the VM controller to take a snapshot.

This presents a problem that we can not revert a VM to the current
snapshot -- which is one of the most common workflows.  To address
that, we will not watch the VirtualMachineSnapshot type custom
resources that are not in ready state and reconcile those from the VM
controller.  Even though this is not ideal (a controller reconciling a
resources for which there's a dedicated controller already exists),
this is necessary to keep all the VM configuration changes in a single
controller so we don't have to use distributed locking between
controllers.

This change:
- Updates the APIs with correct commentary
- Modifies the snapshot controller to always set OwnerRef on the
snapshot pointing to the VM.
- Modifies the snapshot controller to not patch the
spec.currentSnapshot on the VM.
- Modifies the VM controller to enqueue reconcile requests to VMs for
any events on a snapshot resource that is not ready.
- Implement an ordering in the snapshot workflow so that the oldest
snapshot request is processed.
- Modifies the revert workflow to clear out the spec.currentStatus (if
set) so that users can revert the VM to the current snapshot.