Remove references to VM class controller name and predicate handling (#1048)
We introduced VM class controller name when there was a need to have
controllers other than VM operator reconcile VMs.  For example, a
controller written to reconcile VM resources on Fusion.

This code has not been used since its inception, so there's no point
maintaining it.  This change removes it from our codebase.  I am
leaving the controllerName field in the types as is for now.  It can
be deprecated and removed in a separate change.

Additionally, this change removes all the env and configs related to
the VM class controller.