Update how GetVirtualMachine() lookups VMs
Lookup by MoID first. In newer VC it is guaranteed to not ever be
reused and just the easiest thing to reference when set.

Use FindAllByUuid() for the instanceUUID lookup, and return an error
if there are multiple VMs. The VM config InstanceUUID generally is but
not guaranteed to always be unique. When there are multiple matches,
FindByUuid() does not return an error but just one of the results,
which means just one of the VMs (and it isn't clear if that will
always be the same one).

Fix Instance UUID lookup: in a55bfa99 we added VM Spec.InstanceUUID
but did not update the lookup done in GetVirtualMachine() which was
still using the k8s VM object UID. This means that we were never
actually finding new VMs via the instance UUID.

Remove the BiosUUID lookup. Like InstanceUUID it is not guaranteed
to be unique and doesn't add anything over the InstanceUUID.