VMICache location / diskpromo ignore snap+guest
This patch updates the VMICache controller to create a top-level
directory per cached item. This ensures there is no single, top-level
directory per datastore for all cached disks/files. On VSAN, a
top-level datastore directory is a volume, and having multiple,
concurrent threads read/write to this volume can lead to read/write
lock errors.

This patch also updates the disk promotion reconciler to ignore
any disks that participate in snapshots. This prevents a VM from
constantly being tagged as out-of-sync for disk promotion if the
VM has a snapshot.

Finally, this patch causes disk promotion to wait until there is
not a pending or running guest customization.