Revert "🌱 Update vm.status.currentSnapshot in snapshot controller (#1043)" (#1060)
This reverts commit 70832007692938ff5789ac79f1ca395f052dcbb5.

This essentially broke snapshot creation workflow because it removed the code which modifies the spec.currentSnapshot in the create snapshot workflow. This was because we were going to drive the snapshot creation through the VMSnapshot custom resource. That work is tracked by #1049. Once that is merged, we can re-introduce parts of this change.