Disable fast deploy for instance storage VMs
For an instance storage VM during fast deploy, we'll likely end
up doing multiple placement calls, because once the IS PVCs are
all bound, we need to do a final placement to get the Datastores.

However, since we're not sending in the selected host to that
later placement call, we may end up with a different host.

That is fixable - at least with PlaceVM - but IS isn't used
very much so defer fixing that until later.